requirements.txt
pandas
numpy
scikit-learn
sentence-transformers
streamlit
openai
google-generativeai


faiss-cpu